refactor(go-backend): replace cmd/server with independent cmd/http and cmd/grpc (#142)
