+++ go-backend/cmd/http/main_test.go
@@ -1,0 +1,8 @@
+package main_test
+
+import "testing"
+
+// TestMain_Compiles is a build verification test.
+func TestMain_Compiles(t *testing.T) {
+	// build verification
+}
